+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - **Breaking**: Updated k8s-ces-setup configuration in vagrant environment and terraform module to support the new structure of the container registry secret. #96
 For terraform variable declaration see `container_registry_secrets` in [variables.tf](terraform/ces-module/variables.tf).
+- Update default setup helm chart version to 3.0.0 in terraform module `ces-module`.
 
 ## [v3.1.0] - 2024-09-19
 ### Changed